// src/models/project_column.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Project columns contain cards of work.
 */
export type project_column = {
    url: string;
    project_url: string;
    cards_url: string;
    /**
     * The unique identifier of the project column
     */
    id: number;
    node_id: string;
    /**
     * Name of the project column
     */
    name: string;
    created_at: string;
    updated_at: string;
};